+++ test/components/theme_menu_options_test.exs
@@ -271,6 +271,112 @@ defmodule PrimerLive.TestComponents.ThemeMenuOptionsTest do
              |> format_html()
   end
 
+  test "Attribute: update_theme_event" do
+    assigns = %{
+      theme_state: %{
+        color_mode: "light",
+        light_theme: "light_high_contrast",
+        dark_theme: "dark_high_contrast"
+      }
+    }
+
+    assert rendered_to_string(~H"""
+           <.theme_menu_options theme_state={@theme_state} update_theme_event="store_browser_settings" />
+           """)
+           |> format_html() ==
+             """
+             <ul class="ActionList" role="listbox">
+             <li class="ActionList-sectionDivider">
+             <h3 class="ActionList-sectionDivider-title">Theme</h3>
+             </li>
+             <li phx-click="store_browser_settings" phx-value-data="light" phx-value-key="color_mode" class="ActionList-item"
+             role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input checked
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Light</span></span></li>
+             <li phx-click="store_browser_settings" phx-value-data="dark" phx-value-key="color_mode" class="ActionList-item"
+             role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Dark</span></span></li>
+             <li phx-click="store_browser_settings" phx-value-data="auto" phx-value-key="color_mode" class="ActionList-item"
+             role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">System</span></span></li>
+             <li class="ActionList-sectionDivider" role="separator" aria-hidden="true"></li>
+             <li class="ActionList-sectionDivider">
+             <h3 class="ActionList-sectionDivider-title">Dark tone</h3>
+             </li>
+             <li phx-click="store_browser_settings" phx-value-data="dark" phx-value-key="dark_theme" class="ActionList-item"
+             role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Dark</span></span></li>
+             <li phx-click="store_browser_settings" phx-value-data="dark_dimmed" phx-value-key="dark_theme"
+             class="ActionList-item" role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Dark dimmed</span></span></li>
+             <li phx-click="store_browser_settings" phx-value-data="dark_high_contrast" phx-value-key="dark_theme"
+             class="ActionList-item" role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input checked
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Dark high contrast</span></span></li>
+             <li phx-click="store_browser_settings" phx-value-data="dark_colorblind" phx-value-key="dark_theme"
+             class="ActionList-item" role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Dark colorblind</span></span></li>
+             <li phx-click="store_browser_settings" phx-value-data="dark_tritanopia" phx-value-key="dark_theme"
+             class="ActionList-item" role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Dark Tritanopia</span></span></li>
+             <li class="ActionList-sectionDivider" role="separator" aria-hidden="true"></li>
+             <li class="ActionList-sectionDivider">
+             <h3 class="ActionList-sectionDivider-title">Light tone</h3>
+             </li>
+             <li phx-click="store_browser_settings" phx-value-data="light" phx-value-key="light_theme" class="ActionList-item"
+             role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Light</span></span></li>
+             <li phx-click="store_browser_settings" phx-value-data="light_high_contrast" phx-value-key="light_theme"
+             class="ActionList-item" role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input checked
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Light high contrast</span></span></li>
+             <li phx-click="store_browser_settings" phx-value-data="light_colorblind" phx-value-key="light_theme"
+             class="ActionList-item" role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Light colorblind</span></span></li>
+             <li phx-click="store_browser_settings" phx-value-data="light_tritanopia" phx-value-key="light_theme"
+             class="ActionList-item" role="option"><span class="ActionList-content"><span
+                class="ActionList-item-visual ActionList-item-visual--leading"><span
+                    class="FormControl-checkbox-wrap ActionList-item-singleSelectCheckmark"><input
+                        class="FormControl-checkbox" type="checkbox" value="true" /></span></span><span
+                class="ActionList-item-label">Light Tritanopia</span></span></li>
+             <li class="ActionList-sectionDivider" role="separator" aria-hidden="true"></li>
+             <li phx-click="store_browser_settings" phx-value-data="" phx-value-key="reset" class="ActionList-item"><span
+             class="ActionList-content"><span class="ActionList-item-label">Reset to default</span></span></li>
+             </ul>
+             """
+             |> format_html()
+  end
+
   test "Attribute: labels" do
     assigns = %{
       theme_state: %{